requirements.txt
seaborn==0.13.0
pandas==2.1.3
matplotlib==3.8.1
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
wordcloud==1.9.2
textblob==0.17.1
geopandas==0.14.1
pillow==10.1.0
networkx==3.2.1
langchain==0.0.335
textblob==0.17.1
openai==1.3.3
sentence-transformers==2.2.2
plotly-express==0.4.1